edubest-backend-final/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.1
django-environ==0.11.2
psycopg[binary]==3.3.3
dj-database-url==2.1.0
djangorestframework-simplejwt==5.3.1
redis==5.0.3
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.4.0
drf-spectacular==0.27.1
python-slugify==8.0.4
django-phonenumber-field==7.3.0
phonenumbers==8.13.30
openpyxl==3.1.2
django-structlog==8.0.0
